// src/app/interfaces/servicio.ts
export interface Servicio {
    idservicio: string;
    fechaCreacion: string;
    idusuario: string;
    idcliente: string;
    codigoArticuloVenta: string;
    codigoArticuloGrafico: string;
    tipo: string;
    descripcion: string;
    fechaEntregaProveedor: string;
    fechaEntregaVenta: string;
    estado: string;
    item: number;

}